+++ src/Haskellorls/Link.hs
@@ -8,11 +8,10 @@ import qualified Data.Text as T
 import qualified Haskellorls.LsColor.Config as Color
 import qualified Haskellorls.NodeInfo as Node
 import qualified Haskellorls.WrappedText as WT
-import qualified System.Posix.Files as Files
 import qualified System.Posix.Types as Types
 
 nodeLinksNumber :: Node.NodeInfo -> Types.LinkCount
-nodeLinksNumber = Files.linkCount . Node.nodeInfoStatus
+nodeLinksNumber = Node.pfsLinkCount . Node.nodeInfoStatus
 
 nodeLinksNumberWithColor :: Color.Config -> Node.NodeInfo -> [WT.WrappedText]
 nodeLinksNumberWithColor config node = [Color.toWrappedText config getter linksNumber]