+++ src/Haskellorls/Recursive/Utils.hs
@@ -6,7 +6,6 @@ where
 
 import qualified Data.Set as S
 import qualified Haskellorls.NodeInfo as Node
-import qualified System.Posix as Files
 import qualified System.Posix.Types as Types
 
 newtype InodeSet = InodeSet (S.Set Types.FileID)
@@ -14,6 +13,6 @@ newtype InodeSet = InodeSet (S.Set Types.FileID)
 excludeAlreadySeenInode :: InodeSet -> [Node.NodeInfo] -> (InodeSet, [Node.NodeInfo])
 excludeAlreadySeenInode (InodeSet inodeSet) nodes = (InodeSet $ S.union inodeSet newInodeSet, newNodes)
   where
-    newInodes = filter (`S.notMember` inodeSet) $ map (Files.fileID . Node.nodeInfoStatus) nodes
+    newInodes = filter (`S.notMember` inodeSet) $ map (Node.pfsFileID . Node.nodeInfoStatus) nodes
     newInodeSet = S.fromList newInodes
-    newNodes = filter ((`S.member` newInodeSet) . Files.fileID . Node.nodeInfoStatus) nodes
+    newNodes = filter ((`S.member` newInodeSet) . Node.pfsFileID . Node.nodeInfoStatus) nodes